Dashboard multiple page update
