--- src/dashapp/DashBoard/static/css/Infection_CSS.py
@@ -1,27 +1,0 @@
-style_title = {
-    'background': '#f2f2f2',
-    'textAlign': 'center',
-    'marginLeft': '100px',
-    'marginRight': '100px'
-}
-style_h4 = {
-    'textAlign': 'center',
-    # 'marginLeft': '100px',
-    # 'marginRight': '100px'
-}
-
-style_checkList = {
-    'textAlign': 'center'
-}
-
-style_div_left = {
-    'float': 'left',
-    'width': '49%',
-    # 'border': '1px solid #F00',
-}
-
-style_div_right= {
-    'float': 'right',
-    'width': '49%',
-    # 'border': '1px solid #F00',
-}
