+++ src/dashapp/DashBoard/static/css/main_page_CSS.py
@@ -1,0 +1,62 @@
+style_title = {
+    'background': '#f2f2f2',
+    'textAlign': 'center',
+    'marginLeft': '100px',
+    'marginRight': '100px'
+}
+
+style_title2 = {
+    'background': '#f2f2f2',
+    'textAlign': 'center',
+    'marginLeft': '100px',
+    'marginRight': '100px'
+}
+
+style_h4 = {
+    'textAlign': 'center',
+    # 'marginLeft': '100px',
+    # 'marginRight': '100px'
+}
+
+style_checkList = {
+    'textAlign': 'center'
+}
+
+style_div_left = {
+    'float': 'left',
+    'width': '48%',
+    'marginLeft': '1%',
+    # 'border': '1px solid #F00',
+}
+
+style_div_right= {
+    'float': 'right',
+    'width': '48%',
+    'marginRight': '1%',
+    # 'border': '1px solid #F00',
+}
+
+style_interactive_button={
+    # # 'flex': 1,
+    # 'float': 'left',
+    'width': '48%',
+    'marginLeft': '1%',
+}
+
+style_interactive_table={
+    # 'float': 'left',
+    'margin-left': '-50px',
+    'margin-top': '-5px'
+}
+
+style_form={
+    'float': 'right'
+}
+
+style_radio={
+    'margin-left': '10px'
+}
+
+external_stylesheets = [
+    'https://codepen.io/chriddyp/pen/bWLwgP.css',
+    'https://codepen.io/chriddyp/pen/brPBPO.css']
