+++ src/dashapp/data/markdown_text_demo
@@ -1,4 +1,4 @@
-#### Simulation Model of Disease Transmission (Markdown)
+#### Simulation Model of Disease Transmission
 ##### Methods:
 - Simulator on community level;
 - Data Collaboration Techniques for privacy preserving;
